Move Common, DataObject to dataset.py, dataobject.py (#1160)
* Remove Common class.

* Rename common.py->dataobject.py

* Remove DataObject class

* Rename Common->DataSet

* Remove unused ActiveArrayInfo from dataobject.py

* Remove unused imports.

* Import DataObject. Remove extra line.

* Update documentation for rename of Common->DataSet
