+++ pyvista/core/__init__.py
@@ -1,6 +1,6 @@
 """Core routines."""
 
-from .common import Common, DataObject
+from .dataset import DataSet, DataObject
 from .composite import MultiBlock
 from .datasetattributes import DataSetAttributes
 from .filters import (CompositeFilters, DataSetFilters, PolyDataFilters,
