+++ pyvista/utilities/helpers.py
@@ -133,7 +133,7 @@ def convert_array(arr, name=None, deep=0, array_type=None):
 
 def is_pyvista_dataset(obj):
     """Return True if the Object is a PyVista wrapped dataset."""
-    return isinstance(obj, (pyvista.Common, pyvista.MultiBlock))
+    return isinstance(obj, (pyvista.DataSet, pyvista.MultiBlock))
 
 
 def point_array(mesh, name):
